personal-portal/python-backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
httpx
openai
